vickylib: split mod into entities
