+++ vicky/src/lib/database/entities/mod.rs
@@ -1,0 +1,5 @@
+mod task;
+mod lock;
+
+pub use task::*;
+pub use lock::*;
